requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2==2.9.10
pycodestyle==2.14.0
python-dotenv==1.1.1
python-slugify==8.0.4
SQLAlchemy==2.0.43
text-unidecode==1.3
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
